+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.4.0
+* `Context` renamed to `HtmlWriterContext`
+* `link` and `script` moved to `BaseHtmlContext`
+
 ## 0.3.0
 * Package changed from `dev.scottpierce.html` to `dev.scottpierce.html.writer`
 * Remove style read API and turn it into a pure streaming API.